Properly map all regions of memory.
Currently, the EBDA (Extended BIOS Data Area) isn't mapped at all
and the Video RAM, Video ROM and system BIOS are treated as a single
chunk of memory.

Fix this by trying to locate and mmap the EBDA, and by mapping the
VRAM, the VBIOS and the System BIOS separately.