Remove dynamically generated code from LRMI.
LRMI dynamically generated the return to 32-bit code (int 0xff), which
causes problem with systems using PaX, where memory pages are either
executable or writable, but never both at once.  Fix this by including
this code in a separate section in the object file and mapping it
to the absolute address 0x9000.

Thanks to the PaX team for useful suggestions and discussions in
http://bugs.gentoo.org/show_bug.cgi?id=226107